+++ src/main/java/de/codescape/jira/plugins/scrumpoker/action/ConfigureScrumPokerAction.java
@@ -3,7 +3,7 @@ package de.codescape.jira.plugins.scrumpoker.action;
 import com.atlassian.jira.issue.CustomFieldManager;
 import com.atlassian.jira.issue.fields.CustomField;
 import com.atlassian.jira.web.action.JiraWebActionSupport;
-import de.codescape.jira.plugins.scrumpoker.service.ScrumPokerSettings;
+import de.codescape.jira.plugins.scrumpoker.service.ScrumPokerSettingsService;
 
 import java.util.List;
 
@@ -17,10 +17,11 @@ public class ConfigureScrumPokerAction extends JiraWebActionSupport {
     private static final String PARAM_ACTION = "action";
 
     private final CustomFieldManager customFieldManager;
-    private final ScrumPokerSettings scrumPokerSettings;
+    private final ScrumPokerSettingsService scrumPokerSettingsService;
 
-    public ConfigureScrumPokerAction(ScrumPokerSettings scrumPokerSettings, CustomFieldManager customFieldManager) {
-        this.scrumPokerSettings = scrumPokerSettings;
+    public ConfigureScrumPokerAction(ScrumPokerSettingsService scrumPokerSettingsService,
+                                     CustomFieldManager customFieldManager) {
+        this.scrumPokerSettingsService = scrumPokerSettingsService;
         this.customFieldManager = customFieldManager;
     }
 
@@ -35,7 +36,7 @@ public class ConfigureScrumPokerAction extends JiraWebActionSupport {
      * Current configured id of the story point field in this Jira instance.
      */
     public String getStoryPointFieldId() {
-        return scrumPokerSettings.loadStoryPointFieldId();
+        return scrumPokerSettingsService.loadStoryPointFieldId();
     }
 
     @Override
@@ -43,7 +44,7 @@ public class ConfigureScrumPokerAction extends JiraWebActionSupport {
         String newStoryPointField = getHttpRequest().getParameter(PARAM_STORY_POINT_FIELD);
         String action = getHttpRequest().getParameter(PARAM_ACTION);
         if (action != null && action.equals("save")) {
-            scrumPokerSettings.persistStoryPointFieldId(newStoryPointField);
+            scrumPokerSettingsService.persistStoryPointFieldId(newStoryPointField);
         }
         return "success";
     }