+++ src/test/java/de/codescape/jira/plugins/scrumpoker/service/ScrumPokerSettingsServiceTest.java
@@ -9,12 +9,12 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
-import static de.codescape.jira.plugins.scrumpoker.service.DefaultScrumPokerSettings.SETTINGS_NAMESPACE;
-import static de.codescape.jira.plugins.scrumpoker.service.DefaultScrumPokerSettings.STORY_POINT_FIELD;
+import static de.codescape.jira.plugins.scrumpoker.service.ScrumPokerSettingsService.SETTINGS_NAMESPACE;
+import static de.codescape.jira.plugins.scrumpoker.service.ScrumPokerSettingsService.STORY_POINT_FIELD;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class DefaultScrumPokerSettingsTest {
+public class ScrumPokerSettingsServiceTest {
 
     private static final String NEW_FIELD_NAME = "MY_NEW_FIELD";
 
@@ -25,7 +25,7 @@ public class DefaultScrumPokerSettingsTest {
     private PluginSettingsFactory pluginSettingsFactory;
 
     @InjectMocks
-    private DefaultScrumPokerSettings defaultScrumPokerSettings;
+    private ScrumPokerSettingsService scrumPokerSettingsService;
 
     @Mock
     private PluginSettings pluginSettings;
@@ -33,14 +33,14 @@ public class DefaultScrumPokerSettingsTest {
     @Test
     public void persistingShouldRemoveTheSettingIfNoFieldIsGiven() {
         when(pluginSettingsFactory.createGlobalSettings()).thenReturn(pluginSettings);
-        defaultScrumPokerSettings.persistStoryPointFieldId("");
+        scrumPokerSettingsService.persistStoryPointFieldId("");
         verify(pluginSettings).remove(SETTINGS_NAMESPACE + "." + STORY_POINT_FIELD);
     }
 
     @Test
     public void persistingShouldSaveTheSettingIfFieldIsGiven() {
         when(pluginSettingsFactory.createGlobalSettings()).thenReturn(pluginSettings);
-        defaultScrumPokerSettings.persistStoryPointFieldId(NEW_FIELD_NAME);
+        scrumPokerSettingsService.persistStoryPointFieldId(NEW_FIELD_NAME);
         verify(pluginSettings).put(SETTINGS_NAMESPACE + "." + STORY_POINT_FIELD, NEW_FIELD_NAME);
     }