+++ changelog.md
@@ -6,6 +6,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 
 * refactor: improve code to transform database model into the view model
 * refactor: rename estimation field service and use dependency injection
+* refactor: rename settings service and use dependency injection
 
 ## [3.0.1] - 2018-07-16